feat(shared): rename call utils to response
